paso prueba2 a scraper
